Learn NumberConstraint from IsPattern
